refactor(precinct-scanner): Move getCvrsFromExport function
This function didn't need to be a callback defined in AppRoot for any reason. In fact, it had extra guarding for the no-election-definition case even though it will only be called after configuration. I moved it into the pollworker screen component. More steps towards simplifying AppRoot so I can refactor it without breaking things.
