+++ frontends/precinct-scanner/src/app_root.tsx
@@ -8,7 +8,6 @@ import {
   AdjudicationReasonInfo,
   OptionalElectionDefinition,
   Provider,
-  CastVoteRecord,
   PrecinctId,
   ElectionDefinition,
 } from '@votingworks/types';
@@ -592,21 +591,6 @@ export function AppRoot({
   const scannedBallotCount = scanner?.status.ballotCount ?? 0;
   const canUnconfigure = scanner?.status.canUnconfigure ?? false;
 
-  const getCvrsFromExport = useCallback(async (): Promise<CastVoteRecord[]> => {
-    if (electionDefinition) {
-      const castVoteRecordsString = await scan.getExport();
-
-      const lines = castVoteRecordsString.split('\n');
-      const cvrs = lines.flatMap((line) =>
-        line.length > 0 ? (JSON.parse(line) as CastVoteRecord) : []
-      );
-      // TODO add more validation of the CVR, move the validation code from election-manager to utils
-      return cvrs.filter((cvr) => cvr._precinctId !== undefined);
-    }
-    return [];
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [electionDefinition, scannedBallotCount]);
-
   // Initialize app state
   useEffect(() => {
     async function initializeScanner() {
@@ -763,7 +747,6 @@ export function AppRoot({
           scannedBallotCount={scannedBallotCount}
           isPollsOpen={isPollsOpen}
           togglePollsOpen={togglePollsOpen}
-          getCvrsFromExport={getCvrsFromExport}
           printer={printer}
           hasPrinterAttached={!!printerInfo}
           isLiveMode={!isTestMode}
